feat: 4단계 요구사항 구현
- UserService의 트랜잭션 로직을 별도의 클래스로 추상화
