+++ app/src/main/java/com/techcourse/service/AppUserService.java
@@ -1,0 +1,42 @@
+package com.techcourse.service;
+
+import com.techcourse.dao.UserDao;
+import com.techcourse.dao.UserHistoryDao;
+import com.techcourse.domain.User;
+import com.techcourse.domain.UserHistory;
+
+import java.util.Optional;
+
+
+public class AppUserService implements UserService {
+
+    private final UserDao userDao;
+    private final UserHistoryDao userHistoryDao;
+
+    public AppUserService(final UserDao userDao, final UserHistoryDao userHistoryDao) {
+        this.userDao = userDao;
+        this.userHistoryDao = userHistoryDao;
+    }
+
+    @Override
+    public Optional<User> findById(final long id) {
+        return userDao.findById(id);
+    }
+
+    @Override
+    public void insert(final User user) {
+        userDao.insert(user);
+    }
+
+    @Override
+    public void changePassword(final long id, final String newPassword, final String createBy) {
+        final var user = findById(id);
+        if (user.isPresent()) {
+            final User presentUser = user.get();
+            presentUser.changePassword(newPassword);
+            userDao.update(presentUser);
+            userHistoryDao.log(new UserHistory(presentUser, createBy));
+        }
+
+    }
+}
