+++ app/src/main/java/com/techcourse/service/UserService.java
@@ -1,43 +1,13 @@
 package com.techcourse.service;
 
-import com.techcourse.dao.UserDao;
-import com.techcourse.dao.UserHistoryDao;
 import com.techcourse.domain.User;
-import com.techcourse.domain.UserHistory;
-import org.springframework.transaction.support.TransactionManager;
 
 import java.util.Optional;
 
+public interface UserService {
+    Optional<User> findById(final long id);
 
-public class UserService {
+    void insert(final User user);
 
-    private final UserDao userDao;
-    private final UserHistoryDao userHistoryDao;
-    private final TransactionManager transactionManager;
-
-    public UserService(final UserDao userDao, final UserHistoryDao userHistoryDao, final TransactionManager transactionManager) {
-        this.userDao = userDao;
-        this.userHistoryDao = userHistoryDao;
-        this.transactionManager = transactionManager;
-    }
-
-    public Optional<User> findById(final long id) {
-        return userDao.findById(id);
-    }
-
-    public void insert(final User user) {
-        userDao.insert(user);
-    }
-
-    public void changePassword(final long id, final String newPassword, final String createBy) {
-        transactionManager.executeInTransaction(() -> {
-            final var user = findById(id);
-            if (user.isPresent()) {
-                final User presentUser = user.get();
-                presentUser.changePassword(newPassword);
-                userDao.update(presentUser);
-                userHistoryDao.log(new UserHistory(presentUser, createBy));
-            }
-        });
-    }
+    void changePassword(final long id, final String newPassword, final String createBy);
 }
